added member addition features

+++ src/main/java/com/example/demo/services/TokenService.java
@@ -1,7 +1,5 @@
 package com.example.demo.services;
 
-import java.util.Map;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -28,12 +26,11 @@ public class TokenService {
         String jwtRefreshToken = jwtService.generateRefreshToken(username);
         refreshTokenRepo.save(
                 new RefreshToken(
-                    jwtRefreshToken,
-                    userService.loadUserByUsername(username)
-                )
-        );
+                        jwtRefreshToken,
+                        userService.loadUserByUsername(username)));
         return jwtRefreshToken;
     }
+
     public String generateAccessToken(String refreshToken) {
         String username = jwtService.extractUsername(refreshToken);
         if (refreshTokenRepo.existsByUserUsernameAndTokenAndIsExpiredFalse(username, refreshToken)) {
@@ -48,14 +45,17 @@ public class TokenService {
 
     public boolean isValidRefreshToken(String refreshToken) {
         String username = jwtService.extractUsername(refreshToken);
+        System.out.println(username);
         return refreshTokenRepo.existsByUserUsernameAndTokenAndIsExpiredFalse(username, refreshToken)
-        &&
-        jwtService.validateToken(userService.loadUserByUsername(username), refreshToken)
-        ;
+                &&
+                jwtService.validateToken(userService.loadUserByUsername(username), refreshToken);
 
     }
-    public void invalidateRefreshToken(String refreshToken) {
-        String username = jwtService.extractUsername(refreshToken);
-        refreshTokenRepo.deleteByUserUsername(username);
+
+    public void invalidateRefreshToken(String username) {
+        if (refreshTokenRepo.existsByUserUsername(username)) {
+            refreshTokenRepo.deleteByUserUsername(username);
+            refreshTokenRepo.flush();
+        }
     }
 }
